Moved audio data to be stored in an asset instead of the UAudioManager component.
