+++ Assets/HoloToolkit-Examples/SpatialSound/Audio/ExampleAudioBank.asset.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: e08c94e41fdcaa042af9ed2896f7950a
+timeCreated: 1508414051
+licenseType: Pro
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
