+++ Assets/HoloToolkit-Examples/SpatialSound/Scenes/UAudioManagerTest.unity
@@ -38,7 +38,7 @@ RenderSettings:
   m_ReflectionIntensity: 1
   m_CustomReflection: {fileID: 0}
   m_Sun: {fileID: 0}
-  m_IndirectSpecularColor: {r: 0.44657898, g: 0.4964133, b: 0.5748178, a: 1}
+  m_IndirectSpecularColor: {r: 0.44657868, g: 0.49641263, b: 0.57481706, a: 1}
 --- !u!157 &3
 LightmapSettings:
   m_ObjectHideFlags: 0
@@ -109,6 +109,8 @@ NavMeshSettings:
     manualTileSize: 0
     tileSize: 256
     accuratePlacement: 0
+    debug:
+      m_Flags: 0
   m_NavMeshData: {fileID: 0}
 --- !u!1 &69647937
 GameObject:
@@ -138,247 +140,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 88e5a9c38f6989249b88204932d7884f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  Events:
-  - Name: Laser
-    Spatialization: 0
-    RoomSize: 0
-    MinGain: -96
-    MaxGain: 12
-    AttenuationCurve:
-      serializedVersion: 2
-      m_Curve:
-      - serializedVersion: 2
-        time: 0
-        value: 1
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      - serializedVersion: 2
-        time: 1
-        value: 0
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      m_PreInfinity: 2
-      m_PostInfinity: 2
-      m_RotationOrder: 4
-    SpatialCurve:
-      serializedVersion: 2
-      m_Curve:
-      - serializedVersion: 2
-        time: 0
-        value: 1
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      - serializedVersion: 2
-        time: 1
-        value: 1
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      m_PreInfinity: 2
-      m_PostInfinity: 2
-      m_RotationOrder: 4
-    SpreadCurve:
-      serializedVersion: 2
-      m_Curve:
-      - serializedVersion: 2
-        time: 0
-        value: 0
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      - serializedVersion: 2
-        time: 1
-        value: 0
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      m_PreInfinity: 2
-      m_PostInfinity: 2
-      m_RotationOrder: 4
-    LowPassCurve:
-      serializedVersion: 2
-      m_Curve:
-      - serializedVersion: 2
-        time: 0
-        value: 0
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      - serializedVersion: 2
-        time: 1
-        value: 0
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      m_PreInfinity: 2
-      m_PostInfinity: 2
-      m_RotationOrder: 4
-    ReverbCurve:
-      serializedVersion: 2
-      m_Curve:
-      - serializedVersion: 2
-        time: 0
-        value: 0
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      - serializedVersion: 2
-        time: 1
-        value: 0
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      m_PreInfinity: 2
-      m_PostInfinity: 2
-      m_RotationOrder: 4
-    MaxDistanceAttenuation3D: 100
-    UnityGainDistance: 1
-    AudioBus: {fileID: 0}
-    PitchCenter: 1
-    PitchRandomization: 0
-    VolumeCenter: 1
-    VolumeRandomization: 0
-    PanCenter: 0
-    PanRandomization: 0
-    FadeInTime: 0
-    FadeOutTime: 0
-    InstanceLimit: 0
-    InstanceTimeBuffer: 0
-    AudioEventInstanceBehavior: 0
-    Container:
-      ContainerType: 0
-      Looping: 0
-      LoopTime: 0
-      Sounds:
-      - Sound: {fileID: 8300000, guid: 25f57a423322f5a45bf9490fe9d13300, type: 3}
-        Looping: 0
-        DelayCenter: 0
-        DelayRandomization: 0
-      CrossfadeTime: 0
-      CurrentClip: 0
-  - Name: Vocals
-    Spatialization: 0
-    RoomSize: 0
-    MinGain: -96
-    MaxGain: 12
-    AttenuationCurve:
-      serializedVersion: 2
-      m_Curve:
-      - serializedVersion: 2
-        time: 0
-        value: 1
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      - serializedVersion: 2
-        time: 1
-        value: 0
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      m_PreInfinity: 2
-      m_PostInfinity: 2
-      m_RotationOrder: 4
-    SpatialCurve:
-      serializedVersion: 2
-      m_Curve:
-      - serializedVersion: 2
-        time: 0
-        value: 1
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      - serializedVersion: 2
-        time: 1
-        value: 1
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      m_PreInfinity: 2
-      m_PostInfinity: 2
-      m_RotationOrder: 4
-    SpreadCurve:
-      serializedVersion: 2
-      m_Curve:
-      - serializedVersion: 2
-        time: 0
-        value: 0
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      - serializedVersion: 2
-        time: 1
-        value: 0
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      m_PreInfinity: 2
-      m_PostInfinity: 2
-      m_RotationOrder: 4
-    LowPassCurve:
-      serializedVersion: 2
-      m_Curve:
-      - serializedVersion: 2
-        time: 0
-        value: 0
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      - serializedVersion: 2
-        time: 1
-        value: 0
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      m_PreInfinity: 2
-      m_PostInfinity: 2
-      m_RotationOrder: 4
-    ReverbCurve:
-      serializedVersion: 2
-      m_Curve:
-      - serializedVersion: 2
-        time: 0
-        value: 0
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      - serializedVersion: 2
-        time: 1
-        value: 0
-        inSlope: 0
-        outSlope: 0
-        tangentMode: 0
-      m_PreInfinity: 2
-      m_PostInfinity: 2
-      m_RotationOrder: 4
-    MaxDistanceAttenuation3D: 100
-    UnityGainDistance: 1
-    AudioBus: {fileID: 0}
-    PitchCenter: 1
-    PitchRandomization: 0
-    VolumeCenter: 1
-    VolumeRandomization: 0
-    PanCenter: 0
-    PanRandomization: 0
-    FadeInTime: 0
-    FadeOutTime: 0
-    InstanceLimit: 0
-    InstanceTimeBuffer: 0
-    AudioEventInstanceBehavior: 0
-    Container:
-      ContainerType: 0
-      Looping: 0
-      LoopTime: 0
-      Sounds:
-      - Sound: {fileID: 8300000, guid: aa9275dc39fa76144bcdb5c329400abe, type: 3}
-        Looping: 0
-        DelayCenter: 0
-        DelayRandomization: 0
-      CrossfadeTime: 0
-      CurrentClip: 0
+  DefaultBanks:
+  - {fileID: 11400000, guid: e08c94e41fdcaa042af9ed2896f7950a, type: 2}
+  Events: []
   globalEventInstanceLimit: 0
   globalInstanceBehavior: 0
 --- !u!4 &69647939
