+++ Assets/HoloToolkit/SpatialSound/Scripts/UAudioManager/AudioBank.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: b2a183f0508c00a4bbd7204f52fc2983
+timeCreated: 1508250290
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
