+++ Assets/HoloToolkit/SpatialSound/Scripts/UAudioManager/Editor/AudioEventBankEditor.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 0170452cf5961424a928b17ba71275d5
+timeCreated: 1508253207
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
