+++ Assets/HoloToolkit/SpatialSound/Scripts/UAudioManager/Editor/UAudioManagerEditor.cs
@@ -2,23 +2,31 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using UnityEditor;
+using UnityEngine;
 
 namespace HoloToolkit.Unity
 {
     [CustomEditor(typeof(UAudioManager))]
-    public class UAudioManagerEditor : UAudioManagerBaseEditor<AudioEvent>
+    public class UAudioManagerEditor : UAudioManagerBaseEditor<AudioEvent, AudioEventBank>
     {
+        SerializedProperty PropGlobalEventInstanceLimit;
+        SerializedProperty PropGlobalInstanceBehaviour;
+
         private void OnEnable()
         {
             this.MyTarget = (UAudioManager)target;
             SetUpEditor();
+
+            PropGlobalEventInstanceLimit = this.serializedObject.FindProperty("globalEventInstanceLimit");
+            PropGlobalInstanceBehaviour = this.serializedObject.FindProperty("globalInstanceBehavior");
         }
 
         public override void OnInspectorGUI()
         {
-            EditorGUILayout.PropertyField(this.serializedObject.FindProperty("globalEventInstanceLimit"));
-            EditorGUILayout.PropertyField(this.serializedObject.FindProperty("globalInstanceBehavior"));
-            DrawInspectorGUI(false);
+            EditorGUILayout.PropertyField(PropGlobalEventInstanceLimit);
+            EditorGUILayout.PropertyField(PropGlobalInstanceBehaviour);
+            DrawExportToAsset();
+            DrawBankList();
         }
     }
 }
